improve code portfolio
